pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "guitar-toolkit"
version = "0.1.0"
description = "Guitar practice toolkit: stem separation, BPM detection, synced tab downloads"
requires-python = ">=3.13"
dependencies = [
    "demucs>=4.0",
    "librosa>=0.11",
    "soundfile>=0.13",
    "numpy>=2.0",
]

[project.optional-dependencies]
tabs = ["PyGuitarPro>=0.9"]

[project.scripts]
guitar-demucs = "guitar_toolkit.demucs_mix:main"
guitar-bpm = "guitar_toolkit.bpm:main"
guitar-download = "guitar_toolkit.songs.download:main"
guitar-tabs = "guitar_toolkit.tabs.download:main"
guitar-sync = "guitar_toolkit.tabs.sync:main"
guitar-prepare = "guitar_toolkit.songs.workflow:main"

[tool.setuptools.packages.find]
where = ["src"]
